# backend/api/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação de API da aplicação Intellecta",
        "version": "0.1"
    },
    "paths": {
...
        "/institutions/{institution_id}/users": {
            "get": {
                "tags": [
                    "Usuários da Instituição"
                ],
                "summary": "Obter usuários da instituição",
                "description": "Retorna todos os usuários pertencentes a uma instituição específica",
                "operationId": "getInstitutionUsers",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários da instituição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InstitutionUserDto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Instituição não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/institutions/{institution_id}/users/invite": {
            "post": {
                "tags": [
                    "Usuários da Instituição"
                ],
                "summary": "Convidar usuários para instituição",
                "description": "Envia convites para usuários se juntarem à instituição",
                "operationId": "inviteUsers",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Detalhes do convite",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "convites"
                                ],
                                "properties": {
                                    "convites": {
                                        "description": "Lista de e-mails para convidar",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "usuario@exemplo.com"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Convites enviados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InvitationResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Entrada inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/institutions/{institution_id}/users/{institution_user_id}/change-role": {
            "patch": {
                "tags": [
                    "Usuários da Instituição"
                ],
                "summary": "Alterar cargo do usuário na instituição",
                "description": "Atualiza o cargo de um usuário dentro da instituição (apenas administradores)",
                "operationId": "changeUserRole",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "institution_user_id",
                        "in": "path",
                        "description": "ID do vínculo usuário-instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Novo cargo para o usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_role"
                                ],
                                "properties": {
                                    "new_role": {
                                        "description": "Novo cargo do usuário",
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "teacher",
                                            "student"
                                        ],
                                        "example": "teacher"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cargo alterado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstitutionUserDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - apenas administradores podem alterar cargos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado na instituição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cargo inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao atualizar cargo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/institutions/{institution_id}/users/{institution_user_id}": {
            "delete": {
                "tags": [
                    "Usuários da Instituição"
                ],
                "summary": "Remover usuário da instituição",
                "description": "Remove um usuário da instituição (apenas administradores ou o próprio usuário)",
                "operationId": "removeUser",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "institution_user_id",
                        "in": "path",
                        "description": "ID do vínculo usuário-instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário removido da instituição com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - apenas administradores ou o próprio usuário podem remover",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vínculo usuário-instituição não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao remover usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InstitutionUserDto": {
                "description": "DTO de usuário da instituição",
                "properties": {
                    "institution_user_id": {
                        "description": "ID do relacionamento",
                        "type": "string",
                        "format": "uuid"
                    },
                    "institution_id": {
                        "description": "ID da instituição",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "description": "ID do usuário",
                        "type": "string",
                        "format": "uuid"
                    },
                    "role": {
                        "description": "Cargo do usuário",
                        "type": "string",
                        "enum": [
                            "admin",
                            "teacher",
                            "student"
                        ]
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDto"
                    }
                },
                "type": "object"
            },
            "InvitationResponse": {
                "description": "Resposta de convite",
                "properties": {
                    "invitation_id": {
                        "description": "ID do convite",
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "description": "E-mail convidado",
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "description": "Cargo proposto",
                        "type": "string",
                        "enum": [
                            "admin",
                            "teacher",
                            "student"
                        ]
                    },
                    "token": {
                        "description": "Token de convite",
                        "type": "string"
                    },
                    "expires_at": {
                        "description": "Data de expiração",
                        "type": "string",
                        "format": "date-time"
                    },
                    "institution_id": {
                        "description": "ID da instituição",
                        "type": "string",
                        "format": "uuid"
                    },
                    "invited_by": {
                        "description": "ID do usuário que convidou",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "File": {
                "description": "Represents a file in the system",
                "properties": {
                    "file_id": {
                        "description": "Unique identifier for the file",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "url": {
                        "description": "URL to access the file",
                        "type": "string",
                        "format": "uri",
                        "example": "https://storage.example.com/files/profile.jpg"
                    },
                    "filename": {
                        "description": "Original filename",
                        "type": "string",
                        "maxLength": 255,
                        "example": "profile.jpg"
                    },
                    "mime_type": {
                        "description": "MIME type of the file",
                        "type": "string",
                        "example": "image/jpeg",
                        "nullable": true
                    },
                    "size": {
                        "description": "File size in bytes",
                        "type": "integer",
                        "example": 102400,
                        "nullable": true
                    },
                    "uploaded_at": {
                        "description": "When the file was uploaded",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "file_type": {
                        "$ref": "#/components/schemas/FileType"
                    }
                },
                "type": "object"
            },
            "FileType": {
                "description": "Type classification of files",
                "type": "string",
                "enum": [
                    "image",
                    "document",
                    "video",
                    "audio",
                    "archive",
                    "other"
                ],
                "example": "image"
            },
...
            "UserDto": {
                "description": "Data Transfer Object for user information",
                "properties": {
                    "user_id": {
                        "description": "Unique identifier for the user",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "full_name": {
                        "description": "Full name of the user",
                        "type": "string",
                        "maxLength": 255,
                        "example": "João da Silva"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "created_at": {
                        "description": "When the user account was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "changed_at": {
                        "description": "When the user account was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T13:30:00Z"
                    },
                    "profile_picture": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/File"
                            }
                        ],
                        "nullable": true,
                        "description": "User's profile picture"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}